feat : add undo reservation

+++ app/components/AdminDashboard/Reservations/Reservations.tsx
@@ -4,9 +4,9 @@ import "react-multi-date-picker/styles/colors/teal.css";
 import ReservationsTable, {
   Reservation,
 } from "../ReservationsTable/ReservationsTable";
+import DeleteConfirmDialog from "../../DeleteConfirmDialog/DeleteConfirmDialog";
 
 const Reservations = () => {
-  const [description, setDescription] = useState<string>("");
   const [reservations, setReservations] = useState<Reservation[]>([]);
   const [isDialogOpen, setIsDialogOpen] = useState(false);
   const [selectedReservationId, setSelectedReservationId] = useState<
@@ -25,21 +25,40 @@ const Reservations = () => {
   useEffect(() => {
     fetchReservations();
   }, []);
-  {
-    /* <DeleteConfirmDialog
-    isOpen={isDialogOpen}
-    onClose={() => setIsDialogOpen(false)}
-    onConfirm={() => {
-      if (selectedReservationId) {
-        handleDelete(selectedReservationId);
-        setSelectedReservationId(null);
-        setIsDialogOpen(false);
-      }
-    }}
-  /> */
-  }
 
-  return <ReservationsTable reservations={reservations} />;
+  const handleDelete = async (reservationId: number) => {
+    try {
+      await axiosInstance.delete(`/reservation/undo/${reservationId}`);
+      fetchReservations(); // Refresh the reservations list after deletion
+    } catch (error) {
+      console.error("Error deleting reservation:", error);
+    }
+  };
+
+  const handleDeleteClick = (reservationId: number) => {
+    setSelectedReservationId(reservationId);
+    setIsDialogOpen(true);
+  };
+
+  return (
+    <>
+      <ReservationsTable
+        reservations={reservations}
+        onDeleteClick={handleDeleteClick}
+      />
+      <DeleteConfirmDialog
+        isOpen={isDialogOpen}
+        onClose={() => setIsDialogOpen(false)}
+        onConfirm={() => {
+          if (selectedReservationId) {
+            handleDelete(selectedReservationId);
+            setSelectedReservationId(null);
+            setIsDialogOpen(false);
+          }
+        }}
+      />
+    </>
+  );
 };
 
 export default Reservations;
